feat(convertkit): encapsulate the logic as a hooks
