+++ packages/convertkit/src/types.ts
@@ -1,0 +1,5 @@
+export type ConvertkitSubscriber = {
+  id: string
+  email: string
+  first_name: string
+}
